requirements.txt
# --- Build Tools ---
pip>=25.3
setuptools>=75.0.0
wheel

# --- Core Backend ---
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
starlette>=0.37.0
python-multipart>=0.0.9
click>=8.1.7

# --- Config & Validation ---
pydantic>=2.7.0
pydantic-settings>=2.2.0

# --- Networking & Utilities ---
httpx>=0.27.0
python-dotenv>=1.0.1
rich>=13.7.1

# --- AI / OpenAI (Assistants, Files, Vision, Responses API)
openai>=1.60.0

# --- Image / OCR utils ---
pillow>=10.3.0
reportlab>=4.1.0
pypdf>=5.0.0
pdfplumber>=0.11.4
pytesseract>=0.3.10

# --- Database / Vector / Supabase ---
sqlalchemy>=2.0.28
psycopg[binary]>=3.1.18
psycopg2-binary>=2.9.10
pgvector>=0.2.5
supabase>=2.4.0
